% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcDividend_la.R
\name{CalcDividend_la}
\alias{CalcDividend_la}
\title{Calculate Local Authority Level Dividend}
\usage{
CalcDividend_la(
  profiles = smkfreediv::tobacco_profiles,
  clean_income,
  clean_expenditure,
  upshift = 1.57151042,
  div = 0.93
)
}
\arguments{
\item{profiles}{PHE local tobacco profiles.}

\item{clean_income}{Data table. Cleaned income data.}

\item{clean_expenditure}{Data table. Cleaned local authority level weekly expenditure.}

\item{upshift}{Numeric. Parameter to upshift expenditures calculated from the toolkit data.}

\item{div}{Numeric. Proportion of expenditure paid in tax.}
}
\value{

}
\description{
Calculate the smoke-free dividend for each local authority.
}
\examples{

\dontrun{

}
}
