% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcDividend_la_sim.R
\name{CalcDividend_la_sim}
\alias{CalcDividend_la_sim}
\title{Simulate Local Authority Dividend Calculation}
\usage{
CalcDividend_la_sim(
  data,
  upshift = 1.57151042,
  div = 0.93,
  n_sim = 100,
  seed = 2021
)
}
\arguments{
\item{data}{Data table. Cleaned toolkit data}

\item{upshift}{Numeric. Parameter to upshift expenditures calculated from the toolkit data.}

\item{div}{Numeric. Proportion of expenditure paid in tax.}

\item{n_sim}{Numeric. Number of simulations.}

\item{seed}{Numeric. Random number seed.}
}
\value{

}
\description{
Calculate the smoke-free dividend for each local authority, using simulation
methods to incorporate parameter uncertainty in weekly expenditure, local smoking
prevalence, local authority average incomes.
}
\examples{

\dontrun{

}
}
