% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcWeekSpend.R
\name{CalcWeekSpend}
\alias{CalcWeekSpend}
\title{Weekly Tobacco Spend}
\usage{
CalcWeekSpend(data, strat_vars, upshift = 1)
}
\arguments{
\item{data}{Data table. The cleaned toolkit data.}

\item{strat_vars}{Character vector. Variables over which to calculate the weekly mean expenditure}

\item{upshift}{Numeric. Parameter to upshift expenditures calculated from the toolkit data.}
}
\value{

}
\description{
Generates weekly spending on tobacco as a weighted mean across individuals in the sample.
}
\examples{

\dontrun{

}
}
