% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanToolkit.R
\name{CleanToolkit}
\alias{CleanToolkit}
\title{Clean Toolkit Data}
\usage{
CleanToolkit(data = data, start_month, end_month)
}
\arguments{
\item{data}{Data table. Toolkit data.}

\item{start_month}{Numeric. Toolkit tart month for the analysis period.}

\item{end_month}{Numeric. Toolkit end month for the analysis period.}
}
\value{

}
\description{
Clean the raw toolkit data to produce a dataset that can be used for the
smoke free dividend calculations
}
\examples{

\dontrun{

}
}
