% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpi_tobacco}
\alias{cpi_tobacco}
\title{CPI Index for Tobacco}
\format{

}
\source{
\href{https://www.ons.gov.uk/economy/inflationandpriceindices/timeseries/d7cb/mm23}{Tobacco CPI time series}
}
\usage{
cpi_tobacco
}
\description{
Consumer price index time series for tobacco products.
}
\keyword{datasets}
