% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeflateToolkit.R
\name{DeflateToolkit}
\alias{DeflateToolkit}
\title{Deflate Toolkit Data}
\usage{
DeflateToolkit(
  data = data,
  index = smkfreediv::cpi_tobacco,
  base_month,
  base_year
)
}
\arguments{
\item{data}{Data table. Toolkit data cleaned using the \code{CleanToolkit} function.}

\item{index}{Data table. Tobacco price data to use in the deflation. Default is the ONS tobacco CPI series.}

\item{base_month}{Numeric. Month of the year to deflate prices to.}

\item{base_year}{End month for the analysis period}
}
\description{
Use the ONS time series on cigarette prices to deflate tobacco expenditures
in the toolkit data to a base month and year
}
\examples{

\dontrun{

}
}
