% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IllicitSpend.R
\name{IllicitSpend}
\alias{IllicitSpend}
\title{Calculate Illicit Spending}
\usage{
IllicitSpend(
  price_fm_paid = 56.9,
  price_fm_illicit = 28.5,
  price_ryo_paid = 26,
  price_ryo_illicit = 13,
  tot_legal_spend_fm,
  tot_legal_spend_ryo,
  tax_gap_data = smkfreediv::tax_gap_data
)
}
\arguments{
\item{price_fm_paid}{Numeric. The price per cigarette of factory-made cigarettes.}

\item{price_fm_illicit}{Numeric. The price per cigarette of illicit factory-made cigarettes.}

\item{price_ryo_paid}{Numeric. The price per cigarette of roll-your-own tobacco.}

\item{price_ryo_illicit}{Numeric. The price per cigarette of illicit roll-your-own tobacco.}

\item{tot_legal_spend_fm}{Numeric. Estimate of total spending on duty-paid factory-made cigarettes.}

\item{tot_legal_spend_ryo}{Numeric. Estimate of total spending on duty-paid roll-your-own tobacco.}
}
\value{

}
\description{
Calculates the spending on illicit tobacco sources. First obtain the market share of
illicit tobacco consumption from HMRC tax gap tables for both factory-made and roll-your-own and obtain an
estimate for expenditure on illicit sources based on this percentage and the estimated expenditure
on duty-paid tobacco. These expenditures are then adjusted for the estimated
illicit prices as a fraction of the duty-paid prices (about a half).
}
\details{
Prices are obtained from \href{https://www.ashscotland.org.uk/media/850413/28-calculating-the-cost-of-smoking-june-2021.pdf}{ASH Scotland},
for. The market share (consumption) figures are taken from the \href{https://www.gov.uk/government/statistics/tobacco-tax-gap-estimates}{tobacco tax gap estimates}.
}
\examples{

\dontrun{


}
}
