% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{income}
\alias{income}
\title{Local Authority Income Data.}
\format{

}
\source{
\href{https://www.ons.gov.uk/employmentandlabourmarket/peopleinwork/earningsandworkinghours/datasets/smallareaincomeestimatesformiddlelayersuperoutputareasenglandandwales}{Income Data}
}
\usage{
income
}
\description{
A dataset containing lower tier local authorities income and population counts. Raw data gives
model-based income estimates at Middle layer Super Output Area (MSOA) level, which are aggregated to lower-tier
local authority level by simple mean averaging.
}
\keyword{datasets}
