% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{la_gor_lookup}
\alias{la_gor_lookup}
\title{Local Authority to Region Lookup}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 348 rows and 4 columns.
}
\usage{
la_gor_lookup
}
\description{
A lookup table to match local authority districts to government office region geographies.
}
\keyword{datasets}
