% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{localauthorities}
\alias{localauthorities}
\title{Local Authority Names and Codes.}
\format{

}
\source{
\href{https://data.gov.uk/dataset/6ee49b1e-0f4d-4079-90f4-b626e36d2035/lower-tier-local-authority-to-upper-tier-local-authority-april-2019-lookup-in-england-and-wales}{Local Authority lookup}
}
\usage{
localauthorities
}
\description{
A dataset containing the names and codes of lower tier local authorities
(LTLAs) and their mapping to upper tier local authorities (UTLAs) - the level at
which PHE tobacco profiles are published.
}
\keyword{datasets}
