% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PHE_tobacco_profiles}
\alias{PHE_tobacco_profiles}
\title{PHE Tobacco Profiles.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 151 rows and 8 columns.
}
\source{
\href{https://fingertips.phe.org.uk/profile/tobacco-control/data#page/0}{PHE local tobacco profiles}
}
\usage{
PHE_tobacco_profiles
}
\description{
A dataset containing smoking prevalence figures for 151 upper tier local authorities in
England, produced by Public Health England. Figures include smoking prevalence and
total number of smokers, and standard errors for both.
}
\keyword{datasets}
