% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{price_cigs}
\alias{price_cigs}
\title{Price of Cigarettes}
\format{

}
\source{
\href{https://www.ons.gov.uk/economy/inflationandpriceindices/timeseries/czmp}{ONS Cigarette prices}
}
\usage{
price_cigs
}
\description{
Monthly time series of the price of a pack of 20 cigarettes from the ONS
}
\keyword{datasets}
