% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{price_ryo_ons}
\alias{price_ryo_ons}
\title{Price of Hand-rolled tobacco}
\format{
An object of class \code{numeric} of length 1.
}
\source{
\itemize{
\item \href{https://www.ons.gov.uk/economy/inflationandpriceindices/datasets/consumerpriceindicescpiandretailpricesindexrpiitemindicesandpricequotes}{Consumer price inflation item indices and price quotes - Price quotes: December 2018 edition}
}
}
\usage{
price_ryo_ons
}
\description{
Scalar average price of a 100g pack of handrolled / roll-your own (RYO) tobacco, obtained from
price quotes collected by the ONS, underlying the CPI and RPI price indices.
Price calculated for December 2018. Unweighted mean price of N = 144 valid observations for price
yields a price for RYO estimated at £45.52 per 100g.
}
\keyword{datasets}
