% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prop_smokers_ENG}
\alias{prop_smokers_ENG}
\title{Proportion of English Smokers}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
prop_smokers_ENG
}
\description{
A scalar proportion of total UK smokers that are English. This is used to adjust UK total duty receipts from tobacco
to a figure representative for England, in order to feed into the calculation of estimated total expenditure on tobacco
in England for the upshift factor to apply to the toolkit data. This approach assumes that spending follows similar patterns
across the four UK nations such that expenditures are distributed the same way as are smokers.
}
\details{
The proportion is calculated using data on smoking status collected in the Annual Population Survey (APS) from 2011 to 2018.
The proportion of UK smokers who are English is calculated separately for each year and averaged across the eight years. The
figure is stable at around 82% and statistically not significantly different year on year.
}
\keyword{datasets}
