% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadToolkit.R
\name{ReadToolkit}
\alias{ReadToolkit}
\title{Read Toolkit Data}
\usage{
ReadToolkit(path, data, save = TRUE, name)
}
\arguments{
\item{path}{path to the directory where the raw SPSS data is saved}

\item{data}{name of the data file (defaults to the April 2021 version)}

\item{save}{Logical. If TRUE, saves out a copy of the data in RDS format to the same directory as the raw data.}

\item{name}{Character. Name of the file to export}
}
\value{

}
\description{
Reads in the raw Smoking Toolkit Study (STS) data in SPSS format
and saves to a chosen directory in RDS format as well as returning it as
an R object
}
\examples{

\dontrun{

}
}
