% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmkfreeDiv.R
\name{SmkfreeDiv}
\alias{SmkfreeDiv}
\title{Calculate Smoke-free Dividend}
\usage{
SmkfreeDiv(
  profiles = smkfreediv::PHE_tobacco_profiles,
  expenditure,
  upshift = 1.423,
  div = 0.93,
  illicit_prop = 0.083,
  n_sim = 10,
  seed = 2021
)
}
\arguments{
\item{profiles}{Data table. local tobacco profiles for 151 UTLAs.}

\item{expenditure}{Data table. local authority level mean weekly tobacco expenditure.}

\item{upshift}{Numeric. Upshift factor to apply to mean weekly tobacco expenditure. Default = 1.423}

\item{div}{Numeric. Proportion of licit expenditure paid in tax or industry profit. Default = 0.93}

\item{illicit_prop}{Numeric. The proportion of total tobacco expenditure which is illicit. Default = 0.083}

\item{n_sim}{Numeric. Number of simulation replications}

\item{seed}{Numeric. Random number seed for the simulation}
}
\description{
Calculate the smoke-free dividend at upper tier local authority (UTLA) level based on input. Requires
a data table of UTLA tobacco profiles and a data table of UTLA mean weekly tobacco expenditure calculated
from toolkit data
}
\examples{

\dontrun{

}
}
