% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tobacco_duty_receipts}
\alias{tobacco_duty_receipts}
\title{Tobacco Duty Receipts}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 30 rows and 6 columns.
}
\source{
\href{https://www.gov.uk/government/statistics/tobacco-bulletin}{ONS Tobacco Bulletin}
}
\usage{
tobacco_duty_receipts
}
\description{
Annual time series of the total duty receipts from factory-made cigarettes and hand-rolled tobacco.
Data by calendar year from 1992 to 2020.
}
\keyword{datasets}
