% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tobacco_profiles}
\alias{tobacco_profiles}
\title{PHE Tobacco Profiles.}
\format{

}
\source{
\href{https://fingertips.phe.org.uk/profile/tobacco-control/data#page/0}{PHE local tobacco profiles}
}
\usage{
tobacco_profiles
}
\description{
A dataset containing smoking prevalence figures for upper tier local authorities in
England, produced by Public Health England. Figures include smoking prevalence and
total number of smokers, including 95% confidence intervals and standard errors for both.
}
\keyword{datasets}
