% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TOOLKIT}
\alias{TOOLKIT}
\title{Processed Toolkit Data}
\format{

}
\usage{
TOOLKIT
}
\description{
Toolkit data taken from the April 2014 - February 2020 waves to be used in calculating the
tobacco expenditure by local authority. Expenditures are deflated to December 2018 prices.
}
\details{
Data includes variables on consumption and expenditure of tobacco products, upper tier local authority
and region indicators, and socioeconomic variables including age, sex, social grade, education,
and labour market status.
}
\keyword{datasets}
