% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{utla_gor_lookup}
\alias{utla_gor_lookup}
\title{Upper Tier Local Authority to Region Lookup}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 151 rows and 2 columns.
}
\usage{
utla_gor_lookup
}
\description{
A lookup table to match local authority districts to government office region geographies.
}
\keyword{datasets}
