% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcDividend_gor_sim.R
\name{CalcDividend_gor_sim}
\alias{CalcDividend_gor_sim}
\title{Simulate Government Office Region Dividend Calculation}
\usage{
CalcDividend_gor_sim(
  data,
  upshift = 1.57151042,
  div = 0.93,
  n_sim = 100,
  seed = 2021,
  illicit_prop = 1298/(14307 + 1298),
  income_var = 3
)
}
\arguments{
\item{data}{Data table. Cleaned toolkit data}

\item{upshift}{Numeric. Parameter to upshift expenditures calculated from the toolkit data.}

\item{div}{Numeric. Proportion of expenditure paid in tax.}

\item{n_sim}{Numeric. Number of simulations.}

\item{seed}{Numeric. Random number seed.}

\item{illicit_prop}{Numeric. The proportion of total tobacco expenditure which is illicit.}

\item{income_var}{Integer. Option governing which income variable is used. 1 for net household income,
2 for net household income equivalised, and 3 for net household income equivalised after housing costs.}
}
\value{

}
\description{
Calculate the smoke-free dividend for each local authority, using simulation
methods to incorporate parameter uncertainty in weekly expenditure, local smoking
prevalence, local authority average incomes.
}
\examples{

\dontrun{

}
}
