% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcUpshift.R
\name{CalcUpshift}
\alias{CalcUpshift}
\title{Calculate Upshift Factor}
\usage{
CalcUpshift(
  data = data,
  LCFS = FALSE,
  price_fm = 8.83,
  duty_fm = 228.29,
  avt_fm = 0.165,
  price_ryo = 51.6,
  duty_ryo = 234.65,
  deflate_from = c(12, 2020),
  deflate_to = c(12, 2018),
  receipts_data = smkfreediv::tobacco_duty_receipts,
  prices_data = smkfreediv::price_cigs,
  adjust = FALSE
)
}
\arguments{
\item{data}{Data table. The cleaned toolkit data.}

\item{LCFS}{Logical. If TRUE use the LCFS grossed-up spend estimated by Howard Reed.}

\item{price_fm}{Numeric. Price of factory made cigarettes per pack of 20.}

\item{duty_fm}{Numeric. Duty per 1,000 cigarettes}

\item{avt_fm}{Numeric. Ad-valorem tax rate on factory made cigarettes.}

\item{price_ryo}{Numeric. Price of 100g of hand-rolling tobacco.}

\item{duty_ryo}{Numeric. Duty per 1kg of hand-rolling tobacco.}

\item{deflate_from}{Numeric vector with 2 arguments containing the month and year corresponding to
the price of hand-rolled tobacco.}

\item{deflate_to}{Numeric vector with 2 arguments containing the month and year to which hand-rolled tobacco
prices should be deflated.}

\item{receipts_data}{Data table. Tobacco duty receipts data.}

\item{prices_data}{Data table. Price data for a 20-pack of cigarettes.}

\item{adjust}{Logical. If TRUE adjust total receipts by the proportion of smokers who are English.}
}
\value{

}
\description{
As survey data tends to underestimate expenditure on tobacco products, this function
combines estimated total expenditure in the Smoking Toolkit Study data with estimates
based on average prices and HMRC duty receipts. The ratio of the total HMRC-based estimated
annual spend to the toolkit-based estimate is the upshift factor.
}
\details{
Calculations are based on the 2018/19 tax year. Total annual expenditure is calculated by
estimating the % of total price of a pack of 20 cigarettes/100g of hand-rolling tobacco which is
made up of excise duty and using this to derive total spending from total duty receipts. The
corresponding figure from the Toolkit data is obtained by calculating mean weekly spend per smoker
by local authority, multiplying this by the number of smokers in the tobacco profiles, then scaling this
figure up to an annual figure and summing over local authorities.
}
\examples{

\dontrun{

## Use the function defaults to obtain the upshift factor.

up <- CalcUpshift(data = data)

## Set LCFS = TRUE. This combined with the other defaults returns
## the estimate by Howard Reed

up <- CalcUpshift(data = data,
                  LCFS = TRUE)

}
}
