% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IllicitSpend.R
\name{IllicitSpend}
\alias{IllicitSpend}
\title{Calculate Illicit Spending}
\usage{
IllicitSpend(
  price_fm,
  price_ryo,
  price_fm_ratio = 28.5/56.9,
  price_ryo_ratio = 13/26,
  illicit_data = smkfreediv::tax_gap_data,
  illicit_data_year = 2018
)
}
\arguments{
\item{price_fm}{Numeric. Average price of legally sourced cigarettes.}

\item{price_ryo}{Numeric. Average price of legally sourced HRT/RYO tobacco.}

\item{price_fm_ratio}{Numeric. The illicit cigarette price as a proportion of legal prices.}

\item{price_ryo_ratio}{Numeric. The illicit HRT price as a proportion of legal prices.}

\item{illicit_data}{Data table. Consumption data for illicit sources of tobacco.}

\item{illicit_data_year}{Numeric. Year of illicit consumption data to use.}
}
\value{

}
\description{
Calculates the spending on illicit tobacco sources. First obtain the market share of
illicit tobacco consumption from HMRC tax gap tables for both factory-made and roll-your-own and obtain an
estimate for expenditure on illicit sources based on this percentage and the estimated expenditure
on duty-paid tobacco. These expenditures are then adjusted for the estimated
illicit prices as a fraction of the duty-paid prices (about a half).
}
\details{
Prices are obtained from \href{https://www.ashscotland.org.uk/media/850413/28-calculating-the-cost-of-smoking-june-2021.pdf}{ASH Scotland},
for. The market share (consumption) figures are taken from the \href{https://www.gov.uk/government/statistics/measuring-tax-gaps-tables}{measuring tax gaps tables 2021}.
}
\examples{

\dontrun{


}
}
