% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PHW_tobacco_profiles}
\alias{PHW_tobacco_profiles}
\title{PHW Tobacco Profiles.}
\format{

}
\source{
\href{https://publichealthwales.shinyapps.io/smokinginwales/}{PHW local tobacco profiles}
}
\usage{
PHW_tobacco_profiles
}
\description{
A dataset containing smoking prevalence figures for the 22 local authorities in
Wales, produced by Public Health Wales. Figures include smoking prevalence and
total number of smokers, and standard errors for both.
}
\keyword{datasets}
