% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_counts}
\alias{pop_counts}
\title{Population.}
\format{

}
\source{
\href{https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationestimates/datasets/populationestimatesforukenglandandwalesscotlandandnorthernireland}{ONS Population Data}
}
\usage{
pop_counts
}
\description{
Mid-2019 ONS 18+ population estimates by local authority.
}
\keyword{datasets}
