% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_counts}
\alias{pop_counts}
\title{Population.}
\format{

}
\source{
\href{https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationestimates/datasets/populationestimatesforukenglandandwalesscotlandandnorthernireland}{Population Data by age, gender, local authority}
}
\usage{
pop_counts
}
\description{
Mid-2018 ONS population estimates by age and sex.
}
\keyword{datasets}
