% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tax_gap_data}
\alias{tax_gap_data}
\title{Measuring Tax Gaps Data}
\format{

}
\source{
\itemize{
\item \href{https://www.gov.uk/government/statistics/measuring-tax-gaps/measuring-tax-gaps-2021-edition-tax-gap-estimates-for-2019-to-2020}{Measuring tax gaps 2021 edition - tax gap estimates for 2019 to 2020}
\item \href{https://www.gov.uk/government/statistics/measuring-tax-gaps-tables}{Measuring tax gaps tables 2021}
}
}
\usage{
tax_gap_data
}
\description{
Based on the most recently available data (2016/2017), the breakdown of UK tobacco
consumption by product into duty-paid, illicit, and cross-border consumption.
}
\keyword{datasets}
