% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tax_gap_data}
\alias{tax_gap_data}
\title{Tax Gap Market Share Data}
\format{

}
\source{
\href{https://www.gov.uk/government/statistics/tobacco-tax-gap-estimates}{Tobacco Tax Gap}
}
\usage{
tax_gap_data
}
\description{
Based on the most recently available data (2016/2017), the breakdown of UK tobacco
consumption by product into duty-paid, illicit, and cross-border consumption.
}
\keyword{datasets}
